# Deduplicate findIndexOrEnd by exporting it from Data.ByteString.Internal