Closes #334 

I am open to other suggestions for the section name "Internal indexing".