@noughtmare 2020-12-19
> Does this PR improve any benchmark results? Feel free to add benchmarks if the current suite doesn't cover the issue from #334.

There are no existing benchmarks that test the affected functions as far as I'm aware. I could add a new one in this PR, but then we would only see the new "score". The affected functions are the functions in `D.B.Lazy` that use `findIndexOrEnd` which are: `takeWhile`, `dropWhile`, `group` and `groupBy`. I think `group` is an odd one, because all the others have a custom predicate argument. I think `group` could actually be improved by using a function that uses memchr internally, but that is probably something for another issue. So, I am now thinking of adding extra benchmarks in `BenchAll.hs` for the `takeWhile`, `dropWhile` and `groupBy` functions. 

The `BenchAll.hs` file is a bit strange, at the top it says: "Benchmark all 'Builder' functions.", but I think other people have added benchmarks to this file that are not related to the Builder functions, so it might be okay to put these benchmarks there.

Also, I don't know what the best input would be for these benchmarks. For `dropWhile` and `takeWhile` I could just use a lazy ByteString that is a few chunks large where each chunk is filled with a single byte repeated 4k times so that the condition is never met. But `groupBy` might need some other kind of input.