@Bodigrim 2020-12-19
`BenchAll.hs` is kinda kitchen sink these days, please add benchmarks for `dropWhile` and `takeWhile` there.