@noughtmare 2021-01-11
Sorry for the long wait. I'll try to finish this by tomorrow. I'm thinking of reordering the history so that the benchmarks are added before the changes. That way, you can more easily benchmark before and after the changes. I'm also having trouble with naming things.