@noughtmare 2021-01-11
I've added some benchmarks. The results on my machine:

Old:

```
benchmarked findIndexOrEnd/takeWhile
time                 46.15 μs   (45.60 μs .. 46.68 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 45.63 μs   (45.47 μs .. 45.84 μs)
std dev              646.2 ns   (486.9 ns .. 880.3 ns)

benchmarked findIndexOrEnd/dropWhile
time                 49.15 μs   (47.71 μs .. 51.71 μs)
                     0.993 R²   (0.987 R² .. 0.999 R²)
mean                 48.79 μs   (48.45 μs .. 49.39 μs)
std dev              1.566 μs   (998.0 ns .. 2.573 μs)
variance introduced by outliers: 15% (moderately inflated)

benchmarked findIndexOrEnd/break
time                 45.95 μs   (45.12 μs .. 46.91 μs)
                     0.995 R²   (0.989 R² .. 0.999 R²)
mean                 46.32 μs   (45.91 μs .. 47.23 μs)
std dev              1.890 μs   (1.126 μs .. 2.872 μs)
variance introduced by outliers: 22% (moderately inflated)

benchmarked findIndexOrEnd/group zeroes
time                 6.874 μs   (6.823 μs .. 6.945 μs)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 6.907 μs   (6.881 μs .. 6.939 μs)
std dev              95.20 ns   (73.96 ns .. 132.3 ns)

benchmarked findIndexOrEnd/group zero-one
time                 247.0 μs   (240.3 μs .. 252.3 μs)
                     0.994 R²   (0.987 R² .. 0.999 R²)
mean                 261.5 μs   (252.5 μs .. 318.9 μs)
std dev              54.80 μs   (6.846 μs .. 140.5 μs)
variance introduced by outliers: 88% (severely inflated)

benchmarked findIndexOrEnd/groupBy (>=)
time                 191.1 μs   (188.3 μs .. 194.5 μs)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 189.5 μs   (188.9 μs .. 190.6 μs)
std dev              2.651 μs   (1.623 μs .. 4.523 μs)

benchmarked findIndexOrEnd/groupBy (>)
time                 522.4 μs   (511.1 μs .. 537.2 μs)
                     0.995 R²   (0.988 R² .. 0.999 R²)
mean                 533.0 μs   (527.1 μs .. 547.0 μs)
std dev              29.55 μs   (15.59 μs .. 51.65 μs)
variance introduced by outliers: 33% (moderately inflated)
```

New:

```
benchmarked findIndexOrEnd/takeWhile
time                 35.34 μs   (35.21 μs .. 35.46 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 35.47 μs   (35.40 μs .. 35.56 μs)
std dev              271.9 ns   (208.8 ns .. 362.4 ns)

benchmarked findIndexOrEnd/dropWhile
time                 36.65 μs   (36.34 μs .. 37.25 μs)
                     0.996 R²   (0.989 R² .. 1.000 R²)
mean                 36.55 μs   (36.32 μs .. 37.08 μs)
std dev              1.152 μs   (540.8 ns .. 2.187 μs)
variance introduced by outliers: 15% (moderately inflated)

benchmarked findIndexOrEnd/break
time                 50.95 μs   (50.37 μs .. 51.36 μs)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 51.63 μs   (51.17 μs .. 53.43 μs)
std dev              2.906 μs   (259.5 ns .. 6.173 μs)
variance introduced by outliers: 35% (moderately inflated)

benchmarked findIndexOrEnd/group zeroes
time                 6.283 μs   (6.255 μs .. 6.321 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 6.273 μs   (6.258 μs .. 6.288 μs)
std dev              51.72 ns   (40.16 ns .. 74.04 ns)

benchmarked findIndexOrEnd/group zero-one
time                 248.7 μs   (247.9 μs .. 249.5 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 249.1 μs   (248.7 μs .. 249.6 μs)
std dev              1.467 μs   (1.100 μs .. 1.905 μs)

benchmarked findIndexOrEnd/groupBy (>=)
time                 187.1 μs   (186.0 μs .. 188.0 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 185.3 μs   (184.9 μs .. 185.8 μs)
std dev              1.507 μs   (1.244 μs .. 1.924 μs)

benchmarked findIndexOrEnd/groupBy (>)
time                 508.4 μs   (506.9 μs .. 510.2 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 507.5 μs   (506.6 μs .. 508.5 μs)
std dev              3.491 μs   (2.796 μs .. 4.587 μs)
```