@Bodigrim 2021-01-11
On my machine benchmarks are:

```
findIndexOrEnd/takeWhile                 mean 36.32 μs  ( +- 1.579 μs  )
findIndexOrEnd/dropWhile                 mean 39.87 μs  ( +- 1.647 μs  )
findIndexOrEnd/break                     mean 39.18 μs  ( +- 1.237 μs  )
findIndexOrEnd/group zeroes              mean 4.920 μs  ( +- 138.9 ns  )
findIndexOrEnd/group zero-one            mean 232.8 μs  ( +- 19.99 μs  )
findIndexOrEnd/groupBy (>=)              mean 151.0 μs  ( +- 7.790 μs  )
findIndexOrEnd/groupBy (>)               mean 440.4 μs  ( +- 19.53 μs  )
```

vs.

```
findIndexOrEnd/takeWhile                 mean 29.68 μs  ( +- 728.2 ns  )
findIndexOrEnd/dropWhile                 mean 32.34 μs  ( +- 624.3 ns  )
findIndexOrEnd/break                     mean 32.33 μs  ( +- 488.3 ns  )
findIndexOrEnd/group zeroes              mean 4.733 μs  ( +- 78.60 ns  )
findIndexOrEnd/group zero-one            mean 190.5 μs  ( +- 6.970 μs  )
findIndexOrEnd/groupBy (>=)              mean 128.9 μs  ( +- 3.507 μs  )
findIndexOrEnd/groupBy (>)               mean 380.9 μs  ( +- 18.70 μs  )
```

So performance improves all across the board, and I do not observe any regression for `break`.