@noughtmare 2021-01-11
It occurred to me that these benchmarks will need to be changed if rewrite rules are added that would rewrite `dropWhile (== x)` to something that uses `breakByte` or `spanByte` like the strict module has. Maybe we should pre-emptively change the benchmark?