@Bodigrim 2021-01-11
@noughtmare yes, this is a good idea, let's replace `dropWhile (== 0)` by `dropWhile even` or something similar.